+++ engine/language-server/src/test/scala/org/enso/languageserver/vcsmanager/GitSpec.scala
@@ -211,10 +211,16 @@ class GitSpec extends AnyWordSpecLike with Matchers with Effects {
     "reset to last saved state" in new TestCtx with InitialRepoSetup {
       val fooFile = repoPath.resolve("Foo.enso")
       val barFile = repoPath.resolve("Bar.enso")
+      val bazFile = repoPath.resolve("Baz.enso")
       createStubFile(fooFile) should equal(true)
       Files.write(
         fooFile,
         "file contents".getBytes(StandardCharsets.UTF_8)
+      )
+      createStubFile(bazFile) should equal(true)
+      Files.write(
+        bazFile,
+        "baz file contents".getBytes(StandardCharsets.UTF_8)
       )
       val commitResult = vcs.commit(repoPath, "New files").unsafeRunSync()
       commitResult.isRight shouldBe true
@@ -235,17 +241,23 @@ class GitSpec extends AnyWordSpecLike with Matchers with Effects {
 
       val restoreResult = vcs.restore(repoPath, commitId = None).unsafeRunSync()
       restoreResult.isRight shouldBe true
+      restoreResult.getOrElse(Nil) shouldEqual List(
+        Path.of("Bar.enso"),
+        Path.of("Foo.enso")
+      )
 
       val text2 = Files.readAllLines(fooFile)
       text2.get(0) should equal("file contents")
 
       barFile.toFile should exist // TODO: verify this is the expected logic
+      bazFile.toFile should exist
     }
 
-    "reset to a named saved state" in new TestCtx with InitialRepoSetup {
+    "reset to a named saved state while preserving original line endings" in new TestCtx
+      with InitialRepoSetup {
       val fooFile = repoPath.resolve("Foo.enso")
       createStubFile(fooFile) should equal(true)
-      val text1 = "file contents"
+      val text1 = "file contents\r\nand more\u0000"
       Files.write(
         fooFile,
         text1.getBytes(StandardCharsets.UTF_8)
@@ -254,7 +266,7 @@ class GitSpec extends AnyWordSpecLike with Matchers with Effects {
       commitResult.isRight shouldBe true
       val commitId = commitResult.getOrElse(null).commitId
 
-      val text2 = "different contents"
+      val text2 = "different contents\r\nanother line"
       Files.write(
         fooFile,
         text2.getBytes(StandardCharsets.UTF_8)
@@ -263,15 +275,16 @@ class GitSpec extends AnyWordSpecLike with Matchers with Effects {
       val commitResult2 = vcs.commit(repoPath, "More changes").unsafeRunSync()
       commitResult2.isRight shouldBe true
 
-      val fileText1 = Files.readAllLines(fooFile)
-      fileText1.get(0) should equal("different contents")
+      val fileText1 = Files.readString(fooFile)
+      fileText1 should equal(text2)
 
       val restoreResult =
         vcs.restore(repoPath, Some(commitId)).unsafeRunSync()
       restoreResult.isRight shouldBe true
+      restoreResult.getOrElse(Nil) shouldEqual List(Path.of("Foo.enso"))
 
-      val fileText2 = Files.readAllLines(fooFile)
-      fileText2.get(0) should equal("file contents")
+      val fileText2 = Files.readString(fooFile)
+      fileText2 should equal(text1)
     }
 
     "report problem when named save does not exist" in new TestCtx